+++ forseti/tool.py
@@ -6,7 +6,10 @@ import warnings
 from abc import abstractmethod
 from collections import namedtuple
 
+import sys
+
 from forseti.tools.tool_options import ToolOptionDataType
+from forseti.tools import tool_options
 from . import tools
 from forseti.tools.abstool import AbsTool
 import os
@@ -108,9 +111,9 @@ class ToolsListModel(QtCore.QAbstractTableModel):
     def data(self, index, role=None):
         if index.isValid():
             data = self._data[index.row()]
-            if role == QtCore.Qt.DisplayRole:
+            if role == QtCore.Qt.DisplayRole or role == QtCore.Qt.EditRole:
                 if index.column() == ToolsListModel.NAME:
-                    return data.name.title()
+                    return data.name
                 if index.column() == ToolsListModel.DESCRIPTION:
                     return data.description
                 else:
@@ -187,7 +190,7 @@ class ToolOptionsPairsModel(ToolOptionsModel):
         if Qt_Orientation == QtCore.Qt.Vertical:
             if role == QtCore.Qt.DisplayRole:
                 title = self._data[index].label
-                return str(title).title()
+                return str(title)
         return QtCore.QVariant()
         # return super().headerData(index, Qt_Orientation, role)
 
@@ -223,7 +226,7 @@ class ToolOptionsModel2(ToolOptionsModel):
         if Qt_Orientation == QtCore.Qt.Vertical:
             if role == QtCore.Qt.DisplayRole:
                 title = self._data[index].name
-                return str(title).title()
+                return str(title)
         return QtCore.QVariant()
 
     def setData(self, index, data, role=None):
@@ -233,6 +236,46 @@ class ToolOptionsModel2(ToolOptionsModel):
         self._data[index.row()].data = data
         return True
 
+
+class ToolOptionsModel3(ToolOptionsModel):
+
+    def __init__(self, data: typing.List[tool_options.UserOptionPythonDataType], parent=None) -> None:
+        super().__init__(parent)
+        self._data: typing.List[tool_options.UserOptionPythonDataType] = data
+
+    def data(self, index, role=None):
+        if index.isValid():
+            if role == QtCore.Qt.DisplayRole:
+                data = self._data[index.row()].data
+                if data is not None:
+                    return str(data)
+                else:
+                    return ""
+            if role == QtCore.Qt.EditRole:
+                return self._data[index.row()].data
+            if role == QtCore.Qt.UserRole:
+                return self._data[index.row()]
+        return QtCore.QVariant()
+
+    def get(self):
+        options = dict()
+        for data in self._data:
+            options[data.label_text] = data.data
+        return options
+
+    def headerData(self, index, Qt_Orientation, role=None):
+        if Qt_Orientation == QtCore.Qt.Vertical:
+            if role == QtCore.Qt.DisplayRole:
+                title = self._data[index].label_text
+                return str(title)
+        return QtCore.QVariant()
+
+    def setData(self, index, data, role=None):
+        if not index.isValid():
+            return False
+        existing_data = self._data[index.row()]
+        self._data[index.row()].data = data
+        return True
 
 def available_tools() -> dict:
     """
@@ -252,6 +295,10 @@ def available_tools() -> dict:
                 if issubclass(module_class, AbsTool):
                     located_tools[module_class.name] = module_class
         except ImportError as e:
-            raise ImportError("Unable to load {}. Reason: {}".format(m, e))
+
+            print("Unable to load {}. Reason: {}".format(m, e), file=sys.stderr)
+            # raise ImportError("Unable to load {}. Reason: {}".format(m, e))
+        # except ModuleNotFoundError as e:
+        #     print("Unable to load {}".format(m))
 
     return located_tools